managed_identity/requirements.txt
azure-identity
azure-keyvault-secrets
